fix: add env.example, add i18n to home and signin pages, update readme.md, changed shared folder structure, moved i18n.ts to locales folder
