+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { I18nextProvider } from 'react-i18next';
-import i18n from './i18n';
+import i18n from './locales/i18n';
 import './index.css';
 import App from './App';
 import store from './redux/store';
