+++ src/pages/HomePage/HomePage.tsx
@@ -1,5 +1,9 @@
+import { useTranslation } from 'react-i18next';
+
 function HomePage() {
-  return <div>Home page</div>;
+  const { t } = useTranslation();
+
+  return <div> {t('examplePageText.homePage')} </div>;
 }
 
 export default HomePage;
